+++ src/Management/test/Endpoint.Test/ContentNegotiationTests.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.Configuration;
 using Steeltoe.Common.TestResources;
+using Steeltoe.Management.Endpoint.Actuators.HeapDump;
 using Steeltoe.Management.Endpoint.Actuators.Loggers;
 
 namespace Steeltoe.Management.Endpoint.Test;
@@ -179,6 +180,52 @@ public sealed class ContentNegotiationTests
 
         using HttpClient client = host.GetTestClient();
         HttpResponseMessage response = await client.SendAsync(requestMessage);
+
+        response.StatusCode.Should().Be(HttpStatusCode.OK);
+    }
+
+    [Fact]
+    public async Task Cannot_use_incompatible_accept_header_in_heap_dump_actuator()
+    {
+        var appSettings = new Dictionary<string, string?>(AppSettings)
+        {
+            ["management:endpoints:heapDump:heapDumpType"] = "gcdump"
+        };
+
+        WebApplicationBuilder builder = TestWebApplicationBuilderFactory.Create();
+        builder.Configuration.AddInMemoryCollection(appSettings);
+        builder.Services.AddHeapDumpActuator();
+
+        await using WebApplication host = builder.Build();
+        await host.StartAsync();
+
+        using HttpClient client = host.GetTestClient();
+        using var request = new HttpRequestMessage(HttpMethod.Get, new Uri("http://localhost/actuator/heapdump"));
+        request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
+        HttpResponseMessage response = await client.SendAsync(request);
+
+        response.StatusCode.Should().Be(HttpStatusCode.NotAcceptable);
+    }
+
+    [Fact]
+    public async Task Can_use_compatible_accept_header_in_heap_dump_actuator()
+    {
+        var appSettings = new Dictionary<string, string?>(AppSettings)
+        {
+            ["management:endpoints:heapDump:heapDumpType"] = "gcdump"
+        };
+
+        WebApplicationBuilder builder = TestWebApplicationBuilderFactory.Create();
+        builder.Configuration.AddInMemoryCollection(appSettings);
+        builder.Services.AddHeapDumpActuator();
+
+        await using WebApplication host = builder.Build();
+        await host.StartAsync();
+
+        using HttpClient client = host.GetTestClient();
+        using var request = new HttpRequestMessage(HttpMethod.Get, new Uri("http://localhost/actuator/heapdump"));
+        request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/octet-stream"));
+        HttpResponseMessage response = await client.SendAsync(request);
 
         response.StatusCode.Should().Be(HttpStatusCode.OK);
     }
